完善逻辑

+++ AutoToilet/RBTopItemView.h
@@ -13,9 +13,9 @@
 @property (nonatomic, assign) BOOL isHaveRightLine;
 @property (nonatomic, assign) BOOL isRank; /** 是否是水压级别 */
 
-@property (nonatomic, assign) CGFloat temperature;
+@property (nonatomic, assign) NSInteger temperature;
 
-@property (nonatomic, assign) CGFloat step;
+@property (nonatomic, assign) NSInteger step;
 
 
 - (instancetype)initWithFrame:(CGRect)frame andTitle:(NSString *)title;
